Copy Requirement Changes
This code contains updates that need to be performed on the Copy Requirement Trigger. The changes include an updated Copy Requirement Trigger and Trigger handler. A new test class needs to be created because the current test class references the old trigger and handler.  Once the test class is completed, user testing needs to be performed with Users in the Editorial Team and Users in the Planning Team.